added playlist and browser modules to isolate code for clarity and referenced in main.py
